Revert "remove mail handling"
