+++ backend/config-example.json
@@ -1,1 +1,1 @@
-{"dbPath": "", "token": "long token here", "email": "email@email.com", "password": "password", "guildId": ""}
+{"token": "long token here", "email": "email@email.com", "password": "password"}
