+++ backend/src/discord/nlp.js
