+++ backend/src/site/main.js
@@ -1,4 +1,4 @@
-exports.run = function (database, queueChannel) {
+exports.run = function (database) {
   'use strict';
   const express = require('express');
   const utils = require('../utils');
@@ -277,13 +277,8 @@ exports.run = function (database, queueChannel) {
   app.use(express.static(legacy));
   const legacyQueueProcessor = require('./legacy-queue');
   app.post('/legacy/queue', async (req, res) => {
-    legacyQueueProcessor.process(req, database, queueChannel)
-      .then(() => {
-        res.sendFile(legacy + '/success.html');
-      })
-      .catch(() => {
-        res.status(500).send("error");
-      });
+    legacyQueueProcessor.process(req);
+    res.sendFile(legacy + '/success.html');
   });
 
   app.get('/legacy', (req, res) => {
